+++ src/tr1/game/level.h
@@ -4,3 +4,5 @@
 
 bool Level_Initialise(const GF_LEVEL *level, GF_SEQUENCE_CONTEXT seq_ctx);
 void Level_Load(const GF_LEVEL *level);
+float Level_GetFogStart(void);
+float Level_GetFogEnd(void);